refactor(pam): extract getColumnIndicator into shared data-explorer-types
