+++ frontend/src/pages/pam/PamDataExplorerPage/components/DataExplorerGrid.tsx
@@ -11,6 +11,7 @@ import { DataGrid, useDataGrid } from "@app/components/v3/generic/DataGrid";
 import { Skeleton } from "@app/components/v3/generic/Skeleton";
 
 import type { ColumnInfo, FieldInfo, ForeignKeyInfo, TableDetail } from "../data-explorer-types";
+import { getColumnIndicator } from "../data-explorer-types";
 import type { FilterCondition, SortCondition } from "../sql-generation";
 import {
   buildCountQuery,
@@ -116,23 +117,6 @@ const SELECT_COLUMN: ColumnDef<RowData> = {
   enablePinning: false,
   enableResizing: false
 };
-
-function getColumnIndicator(
-  colName: string,
-  primaryKeys: string[],
-  fkMap: Map<string, ForeignKeyInfo>
-): { type: "pk" | "fk"; tooltip?: string } | undefined {
-  if (primaryKeys.includes(colName)) return { type: "pk" };
-  const fk = fkMap.get(colName);
-  if (fk) {
-    const targetCol = fk.targetColumns[fk.columns.indexOf(colName)] ?? fk.targetColumns[0];
-    return {
-      type: "fk",
-      tooltip: `\u2192 ${fk.targetSchema}.${fk.targetTable}(${targetCol})`
-    };
-  }
-  return undefined;
-}
 
 function buildColumnDefs(
   cols: ColumnInfo[],
